+++ test/e2e/framework/node/BUILD
@@ -19,7 +19,6 @@ go_library(
         "//staging/src/k8s.io/apimachinery/pkg/util/sets:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/wait:go_default_library",
         "//staging/src/k8s.io/client-go/kubernetes:go_default_library",
-        "//staging/src/k8s.io/client-go/rest:go_default_library",
         "//test/e2e/framework/log:go_default_library",
         "//test/utils:go_default_library",
     ],
